initial restructure
