+++ controllers/providers/aws/predicates.go
@@ -37,6 +37,18 @@ func IsUsingLaunchTemplate(group *autoscaling.Group) bool {
 	return false
 }
 
+type ManagedNodeGroupReconcileState struct {
+	OngoingState             bool
+	FiniteState              bool
+	UnrecoverableError       bool
+	UnrecoverableDeleteError bool
+}
+
+var ManagedNodeGroupOngoingState = ManagedNodeGroupReconcileState{OngoingState: true}
+var ManagedNodeGroupFiniteState = ManagedNodeGroupReconcileState{FiniteState: true}
+var ManagedNodeGroupUnrecoverableError = ManagedNodeGroupReconcileState{UnrecoverableError: true}
+var ManagedNodeGroupUnrecoverableDeleteError = ManagedNodeGroupReconcileState{UnrecoverableDeleteError: true}
+
 func IsNodeGroupInConditionState(key string, condition string) bool {
 	conditionStates := map[string]ManagedNodeGroupReconcileState{
 		"CREATING":      ManagedNodeGroupOngoingState,
@@ -62,6 +74,22 @@ func IsNodeGroupInConditionState(key string, condition string) bool {
 		return false
 	}
 }
+
+type CloudResourceReconcileState struct {
+	OngoingState             bool
+	FiniteState              bool
+	FiniteDeleted            bool
+	UpdateRecoverableError   bool
+	UnrecoverableError       bool
+	UnrecoverableDeleteError bool
+}
+
+var OngoingState = CloudResourceReconcileState{OngoingState: true}
+var FiniteState = CloudResourceReconcileState{FiniteState: true}
+var FiniteDeleted = CloudResourceReconcileState{FiniteDeleted: true}
+var UpdateRecoverableError = CloudResourceReconcileState{UpdateRecoverableError: true}
+var UnrecoverableError = CloudResourceReconcileState{UnrecoverableError: true}
+var UnrecoverableDeleteError = CloudResourceReconcileState{UnrecoverableDeleteError: true}
 
 func IsProfileInConditionState(key string, condition string) bool {
